Refactor hero into its own component
